feat: allow users to update and delete their own profiles

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'prisma/prisma.service';
 import * as bcrypt from 'bcrypt';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -22,7 +22,11 @@ export class UsersService {
     return this.prisma.user.findMany();
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, user: any) {
+    if (user.role !== 'ADMIN' && user.userId !== id) {
+      throw new ForbiddenException('You can only access your own profile');
+    }
+
     return this.prisma.user.findUnique({ where: { id } });
   }
 
@@ -30,14 +34,20 @@ export class UsersService {
     return this.prisma.user.findUnique({ where: { email } });
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(id: number, updateUserDto: UpdateUserDto, user: any) {
+    if (user.role !== 'ADMIN' && user.userId !== id) {
+      throw new ForbiddenException('You can only update your own profile');
+    }
     return this.prisma.user.update({
       where: { id },
       data: updateUserDto,
     });
   }
 
-  async remove(id: number) {
+  async remove(id: number, user: any) {
+    if (user.role !== 'ADMIN' && user.userId !== id) {
+      throw new ForbiddenException('You can only update your own profile');
+    }
     return this.prisma.user.delete({ where: { id } });
   }
 }
